feat: save local assets with a corresponding subpath
