+++ studio/src/app/services/storage/storage.online.service.tsx
@@ -1,10 +1,13 @@
 import {firebase} from '@firebase/app';
 import '@firebase/storage';
-import {Reference} from '@firebase/storage-types';
+import {Reference, ListResult, ListOptions} from '@firebase/storage-types';
 
 import {take} from 'rxjs/operators';
 
 import {AuthUser} from '../../models/auth/auth.user';
+
+import {Resources} from '../../utils/core/resources';
+
 import {AuthService} from '../auth/auth.service';
 import {ErrorService} from '../core/error/error.service';
 
@@ -65,6 +68,69 @@ export class StorageOnlineService {
         this.errorService.error(err.message);
         resolve();
       }
+    });
+  }
+
+  getFiles(next: string | null, folder: string): Promise<StorageFilesList | null> {
+    return new Promise<StorageFilesList | null>((resolve) => {
+      try {
+        this.authService
+          .watch()
+          .pipe(take(1))
+          .subscribe(async (authUser: AuthUser) => {
+            if (!authUser || !authUser.uid || authUser.uid === '' || authUser.uid === undefined) {
+              this.errorService.error('Not logged in.');
+              resolve(null);
+              return;
+            }
+
+            const ref = firebase.storage().ref(`${authUser.uid}/assets/${folder}/`);
+
+            let options: ListOptions = {
+              maxResults: Resources.Constants.STORAGE.MAX_QUERY_RESULTS
+            };
+
+            if (next) {
+              options.pageToken = next;
+            }
+
+            const results: ListResult = await ref.list(options);
+
+            resolve(this.toStorageFileList(results));
+          });
+      } catch (err) {
+        resolve(null);
+      }
+    });
+  }
+
+  private toStorageFileList(results: ListResult): Promise<StorageFilesList> {
+    return new Promise<StorageFilesList>(async (resolve) => {
+      if (!results || !results.items || results.items.length <= 0) {
+        resolve({
+          items: [],
+          nextPageToken: null
+        });
+        return;
+      }
+
+      const storageFiles: Promise<StorageFile>[] = results.items.map(this.toStorageFile);
+      const items: StorageFile[] = await Promise.all(storageFiles);
+
+      resolve({
+        items: items,
+        nextPageToken: results.nextPageToken
+      });
+    });
+  }
+
+  private toStorageFile(ref: Reference): Promise<StorageFile> {
+    return new Promise<StorageFile>(async (resolve) => {
+      resolve({
+        downloadUrl: await ref.getDownloadURL(),
+        fullPath: ref.fullPath,
+        name: ref.name
+      });
     });
   }
 }
