+++ studio/src/app/utils/core/resources.tsx
@@ -4,6 +4,9 @@ export class Resources {
       DECK: {
         TITLE_MAX_LENGTH: 45,
         DESCRIPTION_MAX_LENGTH: 500
+      },
+      STORAGE: {
+        MAX_QUERY_RESULTS: 20
       }
     };
   }
